+++ packages/react-ui/src/widgets/legend/LegendIcon.js
@@ -4,10 +4,6 @@ import { Box, Grid, makeStyles, Typography } from '@material-ui/core';
 const useStyles = makeStyles((theme) => ({
   legendIcon: {
     alignItems: 'center'
-  },
-  icon: {
-    width: theme.spacing(2),
-    height: theme.spacing(2)
   }
 }));
 
@@ -22,8 +18,8 @@ export default function LegendIcon({ legend }) {
 
   const Icons = labels.map((label, idx) => (
     <Grid key={label} container item className={classes.legendIcon}>
-      <Box mr={1.5}>
-        <img src={icons[idx]} className={classes.icon} alt={icons[idx]} />
+      <Box mr={1.5} width={16} height={16}>
+        <img src={icons[idx]} alt={icons[idx]} />
       </Box>
       <Typography variant='overline'>{label}</Typography>
     </Grid>