+++ packages/react-ui/storybook/stories/widgetsUI/PieWidgetUI.stories.js
@@ -1,10 +1,16 @@
 import React from 'react';
 import PieWidgetUI from '../../../src/widgets/PieWidgetUI';
-import { buildReactPropsAsString } from '../../utils';
 
 const options = {
-  title: 'Widgets UI/PieWidgetUI',
-  component: PieWidgetUI
+  title: 'Custom Components/PieWidgetUI',
+  component: PieWidgetUI,
+  parameters: {
+    docs: {
+      source: {
+        type: 'auto'
+      }
+    }
+  }
 };
 
 export default options;
@@ -19,7 +25,6 @@ const Template = (args) => <PieWidgetUI {...args} />;
 export const Default = Template.bind({});
 const DefaultProps = { data: dataDefault };
 Default.args = DefaultProps;
-Default.parameters = buildReactPropsAsString(DefaultProps, 'PieWidgetUI');
 
 export const CustomColors = Template.bind({});
 const CustomColorsProps = {
@@ -51,7 +56,6 @@ const CustomColorsProps = {
   ]
 };
 CustomColors.args = CustomColorsProps;
-CustomColors.parameters = buildReactPropsAsString(CustomColorsProps, 'PieWidgetUI');
 
 export const SelectedCategories = Template.bind({});
 const SelectedCategoriesProps = {
@@ -66,7 +70,3 @@ const SelectedCategoriesProps = {
   onSelectedCategoriesChange: (categories) => console.log(categories)
 };
 SelectedCategories.args = SelectedCategoriesProps;
-SelectedCategories.parameters = buildReactPropsAsString(
-  SelectedCategoriesProps,
-  'PieWidgetUI'
-);